third_party/WebKit/LayoutTests/animations/animationworklet/animator-animate-expected.txt

CONSOLE MESSAGE: line 36: opacity of target is 0.5
